Bug 1577236 - clang-10: Fix a -Wimplicit-int-float-conversion warning in chromium r=bobowen
Fix:
/var/lib/jenkins/workspace/firefox-clang-lld-last/security/sandbox/chromium/base/time/time.h:796:18: error: implicit conversion from 'long' to 'double' changes value from 922337203
  return value > std::numeric_limits<int64_t>::max()
               ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Upstream code is very different now (uses some saturated_cast):
https://chromium.googlesource.com/chromium/src/base/+/master/time/time.h#846

Differential Revision: https://phabricator.services.mozilla.com/D43786

--HG--
extra : moz-landing-system : lando

+++ security/sandbox/chromium/base/time/time.h
@@ -793,7 +793,7 @@ constexpr TimeDelta TimeDelta::Min() {
 constexpr TimeDelta TimeDelta::FromDouble(double value) {
   // TODO(crbug.com/612601): Use saturated_cast<int64_t>(value) once we sort out
   // the Min() behavior.
-  return value > std::numeric_limits<int64_t>::max()
+  return value > double(std::numeric_limits<int64_t>::max())
              ? Max()
              : value < std::numeric_limits<int64_t>::min()
                    ? Min()
